feat: proper u64 representation — Value::UInt(u64), full 0 to 2^64-1 range
u64 values no longer share i64's representation. Signed types (i8-i64)
use Value::Int(i64), unsigned types (u8-u64) use Value::UInt(u64). Both
occupy 8 bytes. An i8 does not consume 128 bits.

VM changes:
- All integer arithmetic opcodes (add/sub/mul/div/mod/pow/neg) dispatch
  on Value::Int vs Value::UInt at runtime — no new opcodes needed
- Comparison, bitwise, and shift opcodes handle both types
- IntNarrow uses i128 intermediate for range checking, produces Int for
  signed targets and UInt for unsigned targets
- FloatToInt produces UInt for values exceeding i64::MAX
- Added LoadUInt opcode for unsigned literal loading
- checked_pow_u64 for unsigned exponentiation

Compiler changes:
- Token::IntLit, ExprKind::IntLit, TypedExprKind::IntLit widened to
  i128 to hold u64-range literals during compilation
- Codegen emits LoadUInt for unsigned-typed integer literals
- Lexer accepts literals up to u64::MAX (was: i64::MAX)

Docs updated:
- SPEC.md: u64 range corrected to 0 to 2^64-1, storage model documented
- types.md: u64 range corrected, limitation note replaced with storage
  model explanation
- EMBEDDING.md: Value::uint() documented

288 tests passing (+6 new u64 range/arithmetic/overflow tests)

+++ basalt-core/src/ast.rs
@@ -98,7 +98,7 @@ pub struct Expr {
 
 #[derive(Debug, Clone)]
 pub enum ExprKind {
-    IntLit(i64),
+    IntLit(i128),
     FloatLit(f64),
     BoolLit(bool),
     StringLit(String),
@@ -237,7 +237,7 @@ pub struct Pattern {
 #[derive(Debug, Clone)]
 pub enum PatternKind {
     Wildcard,
-    IntLit(i64),
+    IntLit(i128),
     FloatLit(f64),
     BoolLit(bool),
     StringLit(String),
